refactor: move some binding files around
